% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDNP.R
\name{CDNP_clusters_sim}
\alias{CDNP_clusters_sim}
\title{Simulate a daily time series of errors based on CDNP clusters}
\usage{
CDNP_clusters_sim(
  get_CDNP_clusters_output,
  simflow,
  USresid = NA,
  initial_resid = 0,
  seed = NA,
  recompute_all_dat = F,
  prevent_neg_flow_after_sample = T,
  simflow_trans = NULL,
  resid_trans = NULL,
  simflow_invtrans = NULL,
  resid_invtrans = NULL,
  sample_option = 1,
  ...
)
}
\arguments{
\item{get_CDNP_clusters_output}{output from get_CDNP_clusters function}

\item{simflow}{vector; streamflow}

\item{initial_resid}{numeric; initial residual error value (default=0)}

\item{seed}{numeric; optional seed}

\item{sample_option}{sample_option=1 (default) samples the residual error to be subtracted from the simulated flow,
sample_option=2 samples the observed flow and determines what the residual error should be based on that,
sample_option=3 samples both the residual error and the observed flow assuming they are both as likely}
}
\value{
vector; predicted errors the same length as simflow
}
\description{
Simulate a daily time series of errors based on CDNP clusters
}
\examples{
library(CDNP)
# first run get_CDNP_clusters function
CDNP_clusters_out<-get_CDNP_clusters(nclusters=10,
                                     ts_data_resid=eg_data$resid,
                                     ts_data_simflow=eg_data$sim)

# perform error simulation
CDNP_clusters_sim_out<-CDNP_clusters_sim(CDNP_clusters_out,eg_data$sim,seed=42)
plot(eg_data$resid,type="l") # original residual
lines(CDNP_clusters_sim_out,col=2,lty=2) # a new simulated residual
}
