% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDNP.R
\name{CDNP_sim}
\alias{CDNP_sim}
\title{Simulate a daily time series of errors based on CDNP lookup table}
\usage{
CDNP_sim(
  get_CDNP_posterior_lookup_output,
  simflow,
  initial_resid = 0,
  seed = NA,
  sampling_method = 2
)
}
\arguments{
\item{get_CDNP_posterior_lookup_output}{output from get_CDNP_posterior_lookup function}

\item{simflow}{vector; streamflow}

\item{initial_resid}{numeric; initial residual error value (default=0)}

\item{seed}{numeric; optional seed}

\item{sampling_method}{sampling_method=1 assumes uniform distribution within bin (with truncation to avoid negative flows),
sampling_method=2 (default) resamples from bin_residuals (excluding those that cause negative flows), sampling_method=3 picks mid value of bin (avoids negatives)}
}
\value{
vector; predicted errors the same length as simflow
}
\description{
Simulate a daily time series of errors based on CDNP lookup table
}
\examples{
library(CDNP)
# first run get_CDNP_clusters function
CDNP_clusters_out<-get_CDNP_clusters(nclusters=10,
                                     nbin=8,
                                     ts_data_resid=eg_data$resid,
                                     ts_data_simflow=eg_data$sim)
# get CNDP lookup
CDNP_posterior_lookup<-get_CDNP_posterior_lookup(CDNP_clusters_out)

# perform error simulation
CDNP_sim_out<-CDNP_sim(CDNP_posterior_lookup,eg_data$sim,seed=42)
plot(eg_data$resid,type="l") # original residual
lines(CDNP_sim_out,col=2,lty=2) # a new simulated residual
}
