% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_alpha_reliability.R
\name{compute_alpha_reliability}
\alias{compute_alpha_reliability}
\title{Calculates the alpha metric from Renard et al. (2010)}
\usage{
compute_alpha_reliability(sim, obs)
}
\arguments{
\item{sim}{matrix; simulated data in different columns}

\item{obs}{vector; observed data, length should be the same as ncol(sim)}
}
\value{
numeric; the alpha value
}
\description{
Calculates the alpha metric from Renard et al. (2010)
}
\examples{
library(CDNP)
set.seed(42)
obs<-rnorm(1000,sd=1)

sd<-1 #  # try with different sd values: 0.5, 1, 2
sim<-NULL
for(i in 1:100){
  sim<-cbind(sim,rnorm(1000,sd=sd))
}
compute_alpha_reliability(sim,obs)
}
