% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pi_resolution.R
\name{compute_pi_abs_resolution}
\alias{compute_pi_abs_resolution}
\title{Calculates the pi abs metric from Renard et al. (2010)}
\usage{
compute_pi_abs_resolution(sim)
}
\arguments{
\item{sim}{matrix; simulated data in different columns}
}
\value{
numeric; the pi abs value
}
\description{
Calculates the pi abs metric from Renard et al. (2010)
}
\examples{
library(CDNP)
set.seed(42)

sd<-1 #  # try with different sd values: 0.5, 1, 2
sim<-NULL
for(i in 1:100){
  sim<-cbind(sim,rnorm(1000,sd=sd))
}
compute_pi_abs_resolution(sim)
}
