% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pi_resolution.R
\name{compute_pi_rel_resolution}
\alias{compute_pi_rel_resolution}
\title{Calculates the pi rel metric from Renard et al. (2010)}
\usage{
compute_pi_rel_resolution(sim)
}
\arguments{
\item{sim}{matrix; simulated data in different columns}
}
\value{
numeric; the pi rel value
}
\description{
Calculates the pi rel metric from Renard et al. (2010)
}
\examples{
library(CDNP)
set.seed(42)

sd<-1 #  # try with different sd values: 0.5, 1, 2
mean<-1 # try with different mean values: 0.5, 1, 2
sim<-NULL
for(i in 1:100){
  sim<-cbind(sim,rnorm(1000,sd=sd,mean=mean))
}
compute_pi_rel_resolution(sim)
}
