% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDNP.R
\name{get_CDNP_posterior_lookup}
\alias{get_CDNP_posterior_lookup}
\title{Create lookup table that determines the conditional probabilities}
\usage{
get_CDNP_posterior_lookup(get_CDNP_clusters_output)
}
\arguments{
\item{get_CDNP_clusters_output}{list; output from get_CDNP_clusters function}
}
\value{
list containing:\cr\cr
\verb{    } posterior_lookup: data.frame; table of conditions and probabilities\cr\cr
\verb{    } bin_residuals: list; contains the residuals for each bin (for resampling)\cr\cr
\verb{    } normalise_data: logical; whether the residual and streamflow are normalised
\verb{    } prevresid_norm_dat: normalised previous day residual data
\verb{    } simflow_norm_dat: normalised streamflow data
}
\description{
Create lookup table that determines the conditional probabilities
}
\examples{
library(CDNP)
# first run get_CDNP_clusters function
CDNP_clusters_out<-get_CDNP_clusters(nclusters=10,
                                     nbin=8,
                                     ts_data_resid=eg_data$resid,
                                     ts_data_simflow=eg_data$sim)
# get CNDP lookup
CDNP_posterior_lookup<-get_CDNP_posterior_lookup(CDNP_clusters_out)
CDNP_posterior_lookup
}
