% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDNP.R
\name{get_optimal_ncluster_and_nbins}
\alias{get_optimal_ncluster_and_nbins}
\title{Uses Differential Evolution optimiser (DEoptim) to determine the number of clusters and bins that give the most reliable results}
\usage{
get_optimal_ncluster_and_nbins(
  resid,
  simflow,
  warmup = 1095,
  nrep = 10,
  itermax = 50,
  lower = c(2, 2),
  upper = c(100, 100),
  sampling_method = 2,
  optim_method = 4
)
}
\arguments{
\item{resid}{vector; residual series}

\item{simflow}{vector; streamflow series}

\item{warmup}{integer; length of data that will be exluded at the beginning of each split period (default=1095)}

\item{nrep}{integer; number of residual simulation replicates (default=10)}

\item{itermax}{integer; maximum number of iterations for DEoptim (default=50)}

\item{lower}{vector of integers; lower bounds for ncluster and nbin (default=c(2,2))}

\item{upper}{vector of integers; upper bounds for ncluster and nbin (default=c(100,100))}
}
\value{
DEoptim object; optimal ncluster and nbin are in $optim$bestmem
}
\description{
This splits the input series into two halves. The first half is used to compute the clusters
and lookup tables. These are used to simulate residuals for the second half and these are
checked against the actual residuals of the second using the  reliability metric (alpha) from
Renard et al. (2010).
}
\examples{
library(CDNP)
optimal_ncluster_and_nbins_out<-get_optimal_ncluster_and_nbins(resid=eg_data$resid,
                                     simflow=eg_data$sim,warmup=30,nrep=2,itermax=3)
# Best member (ncluster and nbin)
optimal_ncluster_and_nbins_out$optim$bestmem

}
\references{
Renard, B., Kavetski, D., Kuczera, G., Thyer, M., & Franks, S. W. (2010). Understanding predictive uncertainty in hydrologic modeling: The challenge of identifying input and structural errors. Water Resources Research, 46(5).
}
