% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDNP.R
\name{normalise}
\alias{normalise}
\title{Normalise values to fit between 0 and 1}
\usage{
normalise(vals, min_vals = NA, max_vals = NA)
}
\arguments{
\item{vals}{numeric values to be normalised}

\item{min_vals}{numeric value representing the minimum value of the population}

\item{max_vals}{numeric value representing the maximum value of the population}
}
\value{
list containing normalised values, min_vals and max_vals
}
\description{
Normalise values to fit between 0 and 1
}
\examples{
library(CDNP)
norm_dat<-normalise(eg_data$resid)
norm_dat
}
