% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDNP.R
\name{predict_kmeans}
\alias{predict_kmeans}
\title{Predict k-means cluster for a new point based on smallest Euclidean distance to cluster centroids}
\usage{
predict_kmeans(new_data, kmeans_model)
}
\arguments{
\item{new_data}{data.frame; first column: new error at t-1, second column: new streamflow at t}

\item{kmeans_model}{output from kmeans function, should be given from get_CDNP_clusters function}
}
\value{
predicted cluster numbers that the new data belongs to
}
\description{
Predict k-means cluster for a new point based on smallest Euclidean distance to cluster centroids
}
\examples{
library(CDNP)
# first run get_CDNP_clusters function
CDNP_clusters_out<-get_CDNP_clusters(nclusters=10,
                                     nbin=8,
                                     ts_data_resid=eg_data$resid,
                                     ts_data_simflow=eg_data$sim)

kmeans_model<-CDNP_clusters_out$kmeans_model
kmeans_model$centers
# by default the data is normalised.
new_data<-data.frame(ts_data_resid_tminus1=0.7,ts_data_simflow=0.11)

predict_kmeans_out<-predict_kmeans(new_data,kmeans_model)
# predicted cluster should be close to (0.7,0.11)
kmeans_model$centers[predict_kmeans_out,]

}
